==> Mag.too puhas kood/1. Linkide_saamine_selenium.ipynb <==
# Saada Lingid igalt curia lehelt kätte ja kirjutada need csv faili. 
# https://selenium-python.readthedocs.io/getting-started.html
# Koodi muutis sobivaks Risto Hinno

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver import ActionChains
from selenium.webdriver.common.by import By
from bs4 import BeautifulSoup
import urllib
import time

### Clean

driver = webdriver.Chrome(ChromeDriverManager().install())
driver.get('http://curia.europa.eu/juris/documents.jsf?oqp=&for=&mat=or&jge=&td=%3B%3B%3BPUB1%3B%3B%3B%3BORDALL&jur=C%2CT%2CF&etat=clot&page=1&dates=&pcs=Oor&lg=&pro=&nat=or&cit=none%252CC%252CCJ%252CR%252C2008E%252C%252C%252C%252C%252C%252C%252C%252C%252C%252Ctrue%252Cfalse%252Cfalse&language=et&avg=&cid=5403783')
driver.maximize_window()

toolTip = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.XPATH, '//*[@id="listeDocuments"]/div/table/tbody/tr[1]/td[6]/table/tbody/tr/td[2]')))
hov = ActionChains(driver).move_to_element(toolTip)
txt = hov.perform()

tooltip_elem=WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.XPATH, '//*[@id="mainForm:j_id70:0:j_id138:2"]')))
first_link=tooltip_elem.find_element_by_css_selector('a').get_attribute('href')
#go to that page
driver.get(first_link)

links=[]
for i in range(1701):
    #fond doc menu
    a = driver.find_elements_by_class_name("menu_documents")
    #get elemnts with kohtulahendid
    docs=a[0].find_elements_by_css_selector('a')#.get_attribute('href')
    #get links
    for doc in docs:
        links.append(doc.get_attribute('href'))
    # get button to next page
    btns_navi=driver.find_elements_by_class_name("btn_navigation")
    for btn in btns_navi:
    #     print(btn.get_attribute("alt"))
        if btn.get_attribute("alt")=='Järgmine lehekülg':
            btn_next_page=btn

    btn_next_page.click()

len(links), len(set(links))

links[23801]

import pandas as pd

df=pd.DataFrame({"links":links})
df.shape

df.to_csv("ELlingid.csv", index=False)
