==> Mag.too puhas kood/2. Lingid_txtks.ipynb <==
# Teine samm, võta csv fail ja moodusta neist html-d

import urllib.request
from bs4 import BeautifulSoup
from pathlib import Path
import pandas as pd
import time

#loeb urli ja loob HTML-i
def query2soup(url):
    try:
        with urllib.request.urlopen(url) as response:
            html = response.read()
        soup = BeautifulSoup(html, 'html.parser')
        return soup
    except Exception as e:
        print(f'excception {e}')
        print(f'url {url}')

# Vaja on lugeda CSV faili ridade kaupa
df=pd.read_csv("ELlingid.csv") # loeb pandase csv-d

df["links"]= df["links"].str.replace("DA", "ET", case = False) 

#failist rea kaupa http kättesaamine ja nendest moodustada html, mis omakorda kirjutada faili

for ind in df.index:
    if ind>22278 and ind<23803:# ind numbreid on muudetud, et osade kaupa HTML andmeid failidesse panna
        html_soup=query2soup(df.links[ind])
        Path('DataET/'+str(ind)+'kohtulahend.txt').write_text(str(html_soup), encoding='utf-8')
        time.sleep(1)
