==> Mag.too puhas kood/3. Txt_sorteerimine(vahendamine).ipynb <==
# Kolmas samm, eemaldada need txt failid, mis ei sisalda tegelikult kohtulahendit 
#(pole vastavasse keelde tõlgitud või sisu puduub)

import urllib.request, os
from bs4 import BeautifulSoup
from pathlib import Path
import shutil
import os

def query2soup(url):
    with urllib.request.urlopen(url) as response:
        html = response.read()
    soup = BeautifulSoup(html, 'html.parser')
    return soup

#Kirjutada need eraldi kausta need, milles ei sisaldu järgmisi lauseid ET
i=0
while i<=23803:
    text=Path('DataET/'+str(i)+'kohtulahend.txt').read_text(encoding='utf-8')
    if "The document is not available in that language" in text:
        i+=1
    if "This document cannot be found"in text:
        i+=1
    else:
        shutil.copyfile(Path('DataET/'+str(i)+'kohtulahend.txt'), Path('Proov/'+str(i)+'kohtulahend.txt'))
        i+=1

# ET eelmise kausta uuesti läbi laskmine, sest kõiki lauseid ei püüdnud kinni ja 
# kirjutada need uude kausta, kus neid lauseid pole
i=0
while i<=23803:
    if os.path.isfile(Path('Proov/'+str(i)+'kohtulahend.txt')): #kontrolli kas fail eksisteerib
        text=Path('Proov/'+str(i)+'kohtulahend.txt').read_text(encoding='utf-8')
        if "The document is not available in that language." in text:
            i+=1
        elif "This document cannot be found." in text:
            i+=1
        else:
            shutil.copyfile(Path('Proov/'+str(i)+'kohtulahend.txt'), Path('DataET_puhas/'+str(i)+'kohtulahend.txt'))
            i+=1
    else:
        i+=1
