==> Mag.too puhas kood/5. Andmed_csv-ks.ipynb <==
# Andmete sisse lugemine pandase tabeliks, et lihtsustada töötlemist!

import urllib.request, os
from bs4 import BeautifulSoup
import pandas as pd
from pathlib import Path
import os

soup_from_txt = BeautifulSoup(Path('DataET_puhas/5kohtulahend.txt').read_text(encoding='utf-8'), 'html.parser')
soup_from_txt

def sisu_tabeliks(supp):
    abilist=[]
    tag_nimi=''
    Pealirjad_list=''
    Tekst_list=''
    Pealkirjad = [['C06Titre3'], ['C27Niveau6titre2'], ['C34Niveau7titre3'], ['C04Titre1'],   
                  ['C06Titre4'],  ['C24Titre4'], ['C23Titre3'], ['C21Titrenumerote1'], ['C21Titre1'], 
                  ['C28Niveau6titre3'], ['C22Titrenumerote2'], ['C21Niveau5titre1'], ['C07Titre5'], 
                  ['C25Niveau5titre5'],  ['C26Titre6'], ['C25Titrenumerote5'], 
                  ['C35Niveau7titre4'], ['C22Niveau5titre2'], ['C33Niveau7titre2'],  ['C24Titrenumerote4'], 
                  ['C22Titre2'], ['C30Niveau6titre5'], ['C23Niveau5titre3'], ['C26Titrenumerote6'], ['C05Titre2'], 
                  ['C29Niveau6titre4'], ['C31Niveau6titre6'], ['C27Titrenumerote7'],  ['C37Niveau7titre6'],
                  ['Cheading2'], ['Cheading1'], ['Cheading3'], ['C23Titrenumerote3'], ['C25Titre5'], ['C07Titre4'], 
                  ['C24Niveau5titre4'], ['C38Niveau7titre7'], ['C36Niveau7titre5'], ['C05Titre3'],  
                  ['C32Niveau7titre1'], ['C26Niveau6titre1']]
    Tekst = [['15Marge2avectiretlong'], ['C74Alineagrandespacement'], ['C12Tiretlong3'], ['C03Tiretlong'], ['C02RPAlineaAltA'],
             ['Cnull'], ['C27Autre'], ['C39Autre'], ['C10Tiretlong1'], 
             ['C01PointAltN'], ['C50Alineaencadre'], ['C12Marge1avectiretlong'], ['C18Marge3avectiretlong'], ['C15Margeretrait0'],
             ['C20Centreitalique'], ['C38Marge1doubleretrait'], ['C80TableLibre'], ['C16Marge3'], ['C09Pointtypographiqueniv1'], 
             ['C28Pointnumerotemanuel'], ['C39autre'], ['C13Marge2'], ['C11Marge1avecretrait'], 
             ['C32Pointnumerotemanuel'], ['C11Tiretlong2'], ['C01PointnumeroteAltN'], ['C08Question'], ['C17Marge3avecretrait'], ['C09Marge0avecretrait'], ['C10Marge1'], ['C41Marge0doubleretrait'], 
             ['C29Marge0doubleretrait'], ['C14Marge2avecretrait'], ['C80Autre'],
             ['C40Pointnumerotemanuel'], ['C33Questionmarge1'], ['CStyle39Autre11ptCentered'], ['C16Margeretrait1'], 
             ['CStyle39Autre11ptCenteredBefore:6ptAfter:0pt'], ['C19Marge1']]
    for elem in supp.findAll('p'):
        if elem.has_attr('class'):
            #print(elem['class'])
            if elem['class'] in Pealkirjad:
                tag_nimi=elem['class'][0]
                Pealkirjad_list = elem.get_text().replace(u'\xa0', u' ')
                abilist.append([tag_nimi, Pealkirjad_list])
            elif elem['class'] in Tekst:
                Tekst_list = elem.get_text().replace(u'\xa0', u' ')
                tag_nimi=elem['class'][0]
                abilist.append([tag_nimi, Tekst_list])
    df_temp=pd.DataFrame(abilist)
    df_temp.columns = ['Tag', 'Lõigu sisu']
    return df_temp
sisu_tabeliks(soup_from_txt)     

def kohtuasja_nr(supp):
    kohtuasja_nr = supp.find('p', class_='C02AlineaAltA')
    kohtuasja_nr_1=kohtuasja_nr.get_text().replace(u'\xa0', u' ').replace(',','').replace('Kohtuasjas ', '').replace('Liidetud kohtuasjades ', '')
    return kohtuasja_nr_1

kohtuasja_nr(soup_from_txt)

#resolutsioon
def resolutsioon(supp):
    tag=''
    abilist=[]
    resolutsioon1=''
    resolutsioon = supp.findAll('p', class_=["C30Dispositifalinea", "C08Dispositif", "C31Dispositiftiretlong","C32Dispositifmarge1","C36Dispositifmarge2avectiretlong","C39Dispositifmarge1doubleretrait", "C37Dispositifmarge0doubleretrait", "C42Dispositifalinea","C02AlineaAltACharChar","C45Dispositifmarge1avectiretlong", "C44Dispositifmarge1avecretrait","C43Dispositiftiretlong","C33Dispositifmarge1avecretrait", "C40Dispositifmarge0avecretrait","C48DispositifIntroduction","C34Dispositifmarge1avectiretlong","C41DispositifIntroduction"])
    for elem in resolutsioon:
        tag=elem['class'][0]
        resolutsioon1=elem.get_text().replace(u'\xa0', u' ')
        abilist.append([tag, resolutsioon1])
    df_temp=pd.DataFrame(abilist)
    df_temp.columns = ['Tag','Lõigu sisu']
    df_temp['Pealkiri1'] = 'Resolutsioon' 
    return df_temp
    
resolutsioon(soup_from_txt)

#pooled     
def poolte_nimed(supp):
    tag = 'C02AlineaAltA'
    abilist = []
    poolte_nimed = ''
    kogu_algus = supp.findAll('p', class_='C02AlineaAltA')
    poolte_nimed = []
    for elem in kogu_algus:
        if elem.findAll('b'):
            poolte_nimed=elem.get_text()
            abilist.append([tag, poolte_nimed])
    df_temp=pd.DataFrame(abilist)
    df_temp.columns = ['Tag','Lõigu sisu']
    df_temp['Pealkiri1'] = 'Poolte_nimed' 
    return df_temp

poolte_nimed(soup_from_txt)

#kuupäev
def kuupaev(supp):
    kuupaev = supp.findAll('p', class_="C19Centre")
    kuupaev1 = kuupaev[1].get_text().replace('(*)', '').replace(u'\xa0', u' ')
    df_temp=pd.DataFrame(columns=['Lõigu sisu','Pealkiri1'], index=range(1))
    df_temp['Lõigu sisu'] = kuupaev1
    df_temp['Pealkiri1'] = 'Kuupaev'
    df_temp['Tag'] = 'C19Centre'
    return df_temp

kuupaev(soup_from_txt)

#Viited märksõnadele
def viited_marksonale(supp):
    viited = supp.findAll('p', class_='C71Indicateur')
    #print(viited)
    viited1 = viited[0].get_text().replace(u'\xa0', u' ')
    df_temp=pd.DataFrame(columns=['Lõigu sisu','Pealkiri1'], index=range(1))
    df_temp['Lõigu sisu'] = viited1
    df_temp['Pealkiri1'] = 'Viited'
    df_temp['Tag'] = 'C71Indicateur'
    return df_temp

viited_marksonale(soup_from_txt)

#koda
def koda(supp):
    koda = supp.findAll('p', class_=["C19Centre", "C93Titredefinitif"])
    koda1 = koda[0].get_text().replace(',','')
    df_temp=pd.DataFrame(columns=['Lõigu sisu','Pealkiri1'], index=range(1))
    df_temp['Lõigu sisu'] = koda1
    df_temp['Pealkiri1'] = 'Koda'
    df_temp['Tag'] = koda[0]['class'][0]
    return df_temp
koda(soup_from_txt)

#kohtumenetluse keel 
def kohtumenetluse_keel(supp):
    kohtumenetluse_keel = supp.findAll('p', class_=["C42FootnoteLangue", 'C49FootnoteLangue'])
    #print(kohtumenetluse_keel)
    keel = kohtumenetluse_keel[0].get_text().replace(u'\xa0', u' ').replace('* Kohtumenetluse keel: ', '').replace('*      Kohtumenetluse keel: ', '').replace('.', '')
    df_temp=pd.DataFrame(columns=['Lõigu sisu','Pealkiri1'], index=range(1))
    df_temp['Lõigu sisu'] = keel
    df_temp['Pealkiri1'] = 'Keel'
    df_temp['Tag'] = kohtumenetluse_keel[0]['class'][0]
    return df_temp
kohtumenetluse_keel(soup_from_txt)

def kohtulahend(supp):
    df_temp = pd.DataFrame(columns=['Tag','Lõigu sisu','Pealkiri1', 'Kohtuasja_nr', "Kohtuasja_ID"])
    df_temp = df_temp.append(koda(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(kuupaev(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(viited_marksonale(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(poolte_nimed(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(kohtumenetluse_keel(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(sisu_tabeliks(supp), ignore_index=True, sort=True)
    df_temp = df_temp.append(resolutsioon(supp), ignore_index=True, sort=True)
    df_temp['Kohtuasja_nr']=kohtuasja_nr(supp)
    return df_temp#.iloc[1]['Lõigu sisu']
kohtulahend(soup_from_txt)

pohitabel = pd.DataFrame(columns=['Tag','Lõigu sisu','Pealkiri1', 'Kohtuasja_nr', "Kohtuasja_ID"])
a=0
for i in range(23803):#23803):
    df_temp = pd.DataFrame(columns=['Tag','Lõigu sisu','Pealkiri1', 'Kohtuasja_nr', "Kohtuasja_ID"])
    try:
        if os.path.isfile(Path('DataET_puhas/'+str(i)+'kohtulahend.txt')): #kontrolli kas fail eksisteerib
            text=BeautifulSoup(Path('DataET_puhas/'+str(i)+'kohtulahend.txt').read_text(encoding='utf-8'))
            df_temp = df_temp.append(kohtulahend(text))
            df_temp['Kohtuasja_ID']=str(i)
            i+=1
        else:
            i+=1
    except Exception as e:
        a+=1
        print('DataET_puhas/'+str(i)+'kohtulahend.txt')
        print(e)
        print(a)
    pohitabel=pohitabel.append(df_temp)
pohitabel.head()

pohitabel.to_csv('PohitabelET_oige.csv', index=False)
