==> Mag.too puhas kood/6. Filtreerimine_kohtulahendid_hinnang.ipynb <==
import pandas as pd
import numpy as np

data = pd.read_csv("PohitabelET_oige.csv")
data.head(20)

#Vaatame kui palju on unikaalseid kohtuasja nr-d
data.Kohtuasja_ID.nunique() 

#Seame need listiks, et oleks hiljem võimalik võrrelda, kas oleme kõik kohtuasjad läbi käinud, leidmaks algused ja lõpud
unikaalsed_kogu=data.Kohtuasja_ID.unique().tolist()
unikaalsed_kogu1=set(unikaalsed_kogu)

# Need pealkirjad on vajalikud selleks, et sealt algab enamasti EL kohtulahendites kohtu hinnang ehk 
# mida juristid esimesena loevad!
# Seda koodi kasutades ja pealkirjasid juurde otsides allpool olevate koodide abil (koodid kuni kohtukuludeni), leidsin hulga, 
# mille lõigu sisu ei ole tühi, pealkiri on tühi ja lõigu sisu on kohtu hinnang selle erinevas väljenduses.
# Seadsin nende lõikude kohtuasja nr-d ka listi, et võrrelda algsete kohtuasja nr-tega, et
# teaksin millistes lahendites neid sõnu ei leidu ja 
# saaksin edasi puhastades kätte kõik vajalikud pealkirjad, kus hinnang sisaldub.
# Osad pealkirjad on välja kommenteeritud, sest hiljem katsetades selgus, et see annab ebatäpseid tulemusi
# Nt "Kolmas küsimus" võis tähendada mõnes lahendis ka liikmesriigi kohtu küsimust, mida polnud vaja.
# Seega jäeti need esialgselt kõrvale (hiljem on võimalik andmeid täiendavalt juurde puhastada)
frame=data[(data['Lõigu sisu'].isnull()==False)&
           (data.Pealkiri1.isnull())&((data['Lõigu sisu']==' Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']==' Euroopa Kohtu hinnang ') |
                                      (data['Lõigu sisu']=='Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']==' – Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='–       Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']=='2)      Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='2.      Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='2.     Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']==' 2. Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='b)      Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']=='b)     Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']=='B.      Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']=='B –  Euroopa Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='III –  Euroopa Kohtu hinnang') |
                                      (data['Lõigu sisu']==' Kohtu hinnang') |
                                      (data['Lõigu sisu']==' Esimese Astme Kohtu hinnang') |
                                      (data['Lõigu sisu']==' Esimese Astme Kohtu hinnang') | 
                                      (data['Lõigu sisu']==' Avaliku Teenistuse Kohtu hinnang') | 
                                      (data['Lõigu sisu']==' Üldkohtu hinnang') |
                                      (data['Lõigu sisu']==' – Üldkohtu hinnang') |
                                      (data['Lõigu sisu']==' b. Avaliku Teenistuse Kohtu hinnang') |
                                      (data['Lõigu sisu']=='–       Avaliku Teenistuse Kohtu hinnang') | 
                                      (data['Lõigu sisu']=='B –  Esimese Astme Kohtu hinnang') |
                                      (data['Lõigu sisu']==' Euroopa Kohtu vastus') | 
                                      #(data['Lõigu sisu']==' Õiguslik käsitlus') | 
                                      #(data['Lõigu sisu']==' Õiguslik käsitlus ') |
                                      #(data['Lõigu sisu']=='III. Õiguslik käsitlus') |
                                      #(data['Lõigu sisu']=='IV.    Õiguslik käsitlus') |
                                      #(data['Lõigu sisu']=='V.      Õiguslik käsitlus') | 
                                      (data['Lõigu sisu']==' Euroopa Kohtu analüüs') |                     
                                      (data['Lõigu sisu']==' Õiguslik analüüs') | 
                                      #(data['Lõigu sisu']==' Põhikohtuasi ja eelotsuse küsimus') | 
                                      #(data['Lõigu sisu']==' Põhikohtuasi ja eelotsuse küsimused') | 
                                      (data['Lõigu sisu']==' Eelotsuse küsimus') |
                                      (data['Lõigu sisu']==' Eelotsuse küsimus ') |
                                      (data['Lõigu sisu']==' Eelotsuse küsimused') |
                                      (data['Lõigu sisu']==' Eelotsuse küsimused ') | 
                                      (data['Lõigu sisu']==' Eelotsuse küsimuse analüüs') | 
                                      (data['Lõigu sisu']==' Eelotsuse küsimuste analüüs') |
                                      (data['Lõigu sisu']==' Eelotsusetaotluste analüüs') |
                                      (data['Lõigu sisu']=='III. Eelotsuse küsimuste analüüs') |
                                      (data['Lõigu sisu']=='III –  Eelotsuse küsimuste analüüs') |
                                      #(data['Lõigu sisu']==' III – Eelotsuse küsimused') |
                                      #(data['Lõigu sisu']=='III –  Eelotsuse küsimused') |
                                      (data['Lõigu sisu']==' Esimese eelotsuse küsimuse analüüs') )]#|
                                      #(data['Lõigu sisu']==' Küsimused') |
                                      #(data['Lõigu sisu']==' Põhiküsimus') |
                                      #(data['Lõigu sisu']==' 1.     Lähteküsimus') |
                                      #(data['Lõigu sisu']==' Esimene küsimus') |
                                      #(data['Lõigu sisu']==' a)     Esimene küsimus') |
                                      #(data['Lõigu sisu']==' Esimene küsimus ') | 
                                      #(data['Lõigu sisu']==' Esimene väide') | 
                                      #(data['Lõigu sisu']==' Esimene ja teine küsimus') | 
                                      #(data['Lõigu sisu']=='C.      Esimene ja teine küsimus') | 
                                      #(data['Lõigu sisu']==' Kaks esimest eelotsuse küsimust') |
                                      #(data['Lõigu sisu']==' Esimene, kolmas ja neljas küsimus ') |
                                      #(data['Lõigu sisu']==' Esimene, teine ja kolmas küsimus') |
                                      #(data['Lõigu sisu']==' Esimesed kolm küsimust') |
                                      #(data['Lõigu sisu']==' Hagi') | 
                                      #(data['Lõigu sisu']==' Hagi ') | 
                                      #(data['Lõigu sisu']=='III –  Hagi') |
                                      #(data['Lõigu sisu']==' Hagimenetlus Üldkohtus') |
                                      #(data['Lõigu sisu']==' Apellatsioonkaebus') |
                                      #(data['Lõigu sisu']==' Apellatsioonkaebused') |
                                      #(data['Lõigu sisu']==' Apellatsioonkaebused') |
                                      #(data['Lõigu sisu']=='IV –  Apellatsioonkaebus') |
                                      #(data['Lõigu sisu']==' VI – Apellatsioonkaebused') |
                                      #(data['Lõigu sisu']=='V –  Apellatsioonkaebus') |
                                      #(data['Lõigu sisu']==' Tühistamisnõue') | 
                                      #(data['Lõigu sisu']==' Sisu') | 
                                      #(data['Lõigu sisu']==' Sisulised küsimused') | 
                                      #(data['Lõigu sisu']==' Vastuvõetavus') |
                                      #(data['Lõigu sisu']=='III. Vastuvõetavus') |
                                      #(data['Lõigu sisu']==' Eelotsusetaotluse vastuvõetavus') | 
                                      #(data['Lõigu sisu']==' Hagi vastuvõetavus') |
                                      #(data['Lõigu sisu']==' Hagi vastuvõetavus ') |
                                      #(data['Lõigu sisu']==' Kolme esimese eelotsuse küsimuse vastuvõetavus') |
                                      #(data['Lõigu sisu']==' Eelotsuse küsimuste vastuvõetavus') |
                                      #(data['Lõigu sisu']==' Menetlus Euroopa Kohtus') | 
                                      #(data['Lõigu sisu']==' Uuesti läbivaatamine') | 
                                      #(data['Lõigu sisu']==' Menetluse suulise osa uuendamise taotlus') | 
                                      #(data['Lõigu sisu']==' Euroopa Kohtu pädevus') | 
                                      #(data['Lõigu sisu']==' Vaidlustatud otsuse tühistamise nõue') | 
                                      #(data['Lõigu sisu']==' Esialgsed tähelepanekud') | 
                                      #(data['Lõigu sisu']==' Esialgsed märkused') | 
                                      #(data['Lõigu sisu']==' Kiirmenetlus') |
                                      #(data['Lõigu sisu']==' Uurimise alustamise otsus') | 
                                      #(data['Lõigu sisu']=='II –  Euroopa Kohtu pädevus') |
                                      #(data['Lõigu sisu']==' Suulise menetluse uuendamise taotlus') | 
                                      #(data['Lõigu sisu']==' Prantsuse valitsuse teisesed nõuded') | 
                                      #(data['Lõigu sisu']==' Direktiivi artikli 7 lõiget 1 käsitlevad küsimused') | 
                                      #(data['Lõigu sisu']==' Direktiivi 93/83 kohaldatavus') | 
                                      #(data['Lõigu sisu']==' EMÜ asutamislepingu artikli 52 tõlgendamine') | 
                                      #(data['Lõigu sisu']==' Vaidlusese') | 
                                      #(data['Lõigu sisu']==' Vaidlustatud määruse tühistamise nõue') | 
                                      #(data['Lõigu sisu']==' Liikmesriigi kohustuste rikkumine') | 
                                      #(data['Lõigu sisu']==' Taganemisõigust käsitlevate direktiivi sätete tingimusteta ja piisavalt täpne olemus') | 
                                      #(data['Lõigu sisu']==' Riigivastutus liikmesriigi seadusandja tegude ja tegevusetuse eest, mis on vastuolus ühenduse õigusega (nii kohtuasja C\xad46/93 kui ka C-48/93 esimene küsimus)') | 
                                      #(data['Lõigu sisu']==' Esimene küsimus (artikli 119 vahetu õigusmõju)') | 
                                      #(data['Lõigu sisu']==' Õigus tõhusale kohtulikule kaitsele') | 
                                      #(data['Lõigu sisu']==' Töötajate õigusi määratlevate direktiivi sätete vahetu õigusmõju') |
                                      #(data['Lõigu sisu']==' Menetlustoimingute taotlus') | 
                                      #(data['Lõigu sisu']==' Ühenduse määrusel põhineva siseriikliku õigusakti täitmise peatamine') |
                                      #(data['Lõigu sisu']==' Esimene väide, et on rikutud määruse nr 1306/2013 artiklit 52 ja määruse nr 885/2006 artikli 11 lõiget 1'))] 
unikaalsed_IDd=frame.Kohtuasja_ID.unique().tolist()
unikaalsed_ID_d_filter=set(unikaalsed_IDd)
frame

# Selle koodiga vaatame andmeid (kohtuasja ID järgi ja mille Pealkirju pole), et 
# saada kätte see koht, kust algab kohtu hinnang, mis oli erinevalt sõnastatud
data[(data.Kohtuasja_ID==22300) & (data.Pealkiri1.isnull())].tail(60)
#Mõned näited, mille all kohtu hinnang oli:
# ID=0 - Eelotsuse küsimuse analüüs; C04Titre1; järgneb C01PointnumeroteAltN; C04Titre1 on eelnevalt rohkem
# 1 - Eelotsuse küsimuste analüüs; C05Titre2; järgneb C01PointnumeroteAltN; C05Titre2 on eelnevalt rohkem
# 2 - Eelotsuse küsimuste analüüs; C04Titre1; C01PointnumeroteAltN;  C04Titre1 on eelnevalt rohkem
# 3 - Eelotsuse küsimuste analüüs; C04Titre1; C01PointnumeroteAltN;
# 4 - Eelotsuse küsimuste analüüs; C04Titre1; C01PointnumeroteAltN;
# 15 - Euroopa Kohtu hinnang; C05Titre2, C01PointnumeroteAltN;
# 16 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 17 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 18 - – Euroopa Kohtu hinnang, C07Titre4, C01PointnumeroteAltN;
# 19 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 20 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 100 - Eelotsuse küsimuse analüüs, C04Titre1, C05Titre2
# 105 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 250 - Eelotsuse küsimuse analüüs, C04Titre1, C05Titre2
# 1000 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 2000 - Eelotsuse küsimuste analüüs,  C04Titre1; C01PointnumeroteAltN;
# 10011 - Eelotsuse küsimus, C04Titre1; C01PointnumeroteAltN;
# 20200 - – Üldkohtu hinnang, C07Titre4; C01PointnumeroteAltN;
# 22120 - Esimese Astme Kohtu hinnang ;  C05Titre2; C01PointnumeroteAltN
#22300 -Esimese Astme Kohtu hinnang; C06Titre3; ; C01PointnumeroteAltN
#Titre-d on liiga erinevad, mistõttu on mõistlik otsida hinnang või analüüs järgi, 
#kuid sisse võib jääda ka alama astme pealkirju, mida võib olla vaja omakorda eemaldada

#Selle koodi abil leidsin, millised kohtuasja nr-d on veel läbi käimata, kus kohtu hinnang võiks olla teisiti nimetatud
len(unikaalsed_kogu1-unikaalsed_ID_d_filter)
# Osad jäid veel alles nt järgmistel põhjustel:
#2764, - C02AlineaAltA all, muidu pole tagitud
# 3010- C02AlineaAltA all, muidu pole tagitud
# 6803- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 7182,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 10415,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 10960,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 11093,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 11571,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 12198,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 12474,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 12813,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 13119,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 13664,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 14912,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 15414,- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!
# 16876, - välja jätta, sest ainult pealkirjad on salvestunud
# 16926,- välja jätta, sest ainult pealkirjad on salvestunud
# 17002,- välja jätta, sest ainult pealkirjad on salvestunud
# 17223,- välja jätta, sest osad andmed pole salvestunud
# 17573,- välja jätta, sest osad andmed pole salvestunud
# 17652,- välja jätta, sest osad andmed pole salvestunud
# 17715,- välja jätta, sest osad andmed pole salvestunud
# 17781,- välja jätta, sest osad andmed pole salvestunud
# 18385,- välja jätta, sest osad andmed pole salvestunud
# 18431,- välja jätta, sest osad andmed pole salvestunud
# 18636,- välja jätta, sest osad andmed pole salvestunud
# 18639,- välja jätta, sest osad andmed pole salvestunud
# 18668,- välja jätta, sest osad andmed pole salvestunud
# 19100- ainult üks C04Titre1- siis pole alampealkirju ja kogu lahend tuleks võtta otsitavaks!

#Kasutades tail/head või loc järgmisi koodiosi otsisin käsitsi õigeid kohtu hinnangu algusele viitavaid sõnu neist
#lahendeist, mis ei tulnud nt "Euroopa Kohtu hinnang" otsinguga välja, sest olid teisiti sõnastatud.
#Selle tulemusena täiendasin alguse sõnade listi
#data[(data.Kohtuasja_ID==0)].head(40)
#data[(data.Kohtuasja_ID==0)].tail(40)
#data[(data.Pealkiri1.isnull())].loc[212880]['Lõigu sisu']


# Osade kohtulahendite hinnangu osa lõppeb kohtukulude määramisega. Seda osa ei ole samamoodi analüüsi vaja.
# Ka see võis olla erinevas vormis kirjas. Seetõttu tuli kõik kohtukulude erinevad vormid üles leida.
frame2=data[(data['Lõigu sisu']==' Kohtukulud') |
            (data['Lõigu sisu']=='Kohtukulud') |
            (data['Lõigu sisu']==' Kohtukulud ') |
            (data['Lõigu sisu']=='  Kohtukulud') |
           (data['Lõigu sisu']=='VII. Kohtukulud') |
           (data['Lõigu sisu']=='IV.    Kohtukulud')|
           (data['Lõigu sisu']=='VI.    Kohtukulud')|
           (data['Lõigu sisu']==' VIII.      Kohtukulud')|
           (data['Lõigu sisu']=='42      Kohtukulud') |
            (data['Lõigu sisu']==' VI.      Kohtukulud') |
            (data['Lõigu sisu']=='IV –  Kohtukulud') |
            (data['Lõigu sisu']=='VI –  Kohtukulud') |
            (data['Lõigu sisu']=='VIII –  Kohtukulud') |
            (data['Lõigu sisu']=='V –  Kohtukulud') |
            (data['Lõigu sisu']=='VII –  Kohtukulud') |
            (data['Lõigu sisu']==' IV – Kohtukulud') |
            (data['Lõigu sisu']=='VI –  Kohtukulud') |
            (data['Lõigu sisu']=='49      Kohtukulud') |
            (data['Lõigu sisu']==' VI     Kohtukulud') |
            (data['Lõigu sisu']==' Kulude kohta')|
            (data['Lõigu sisu']==' VII – Kohtukulud')|
            (data['Lõigu sisu']==' Kulud') |
            (data['Lõigu sisu']=='III –  Kohtukulud') |
            (data['Lõigu sisu']==' 4.     Kohtukulud') |
            (data['Lõigu sisu']==' IV.      Kohtukulud') | 
            (data['Lõigu sisu']=='V.      Kohtukulud') |
            (data['Lõigu sisu']==' V.      Kohtukulud') |
           (data['Lõigu sisu']=='VI – Kohtukulud')]
unikaalsed_lõpud2=set(frame2.Kohtuasja_ID.unique().tolist())

# Selle koodi abil leidsin, millistes lahendites kohtukulud puuduvad. Kasutasin eelnevalt toodud head/tail/loc,
# et vaadelda kohtuotsuste osasid (lõike)
(unikaalsed_kogu1-unikaalsed_lõpud2)
# Neis lahendites, mis siin järele on jäänud, ei ole kohtukulude osa eraldi kirjas.

#Leidnud alguse ja lõpu pealkirjad tuli need eraldi ära märgendada (sildistada), et filtreerida välja ainult sobivad kohtulahendite osad
alguse_tagid=set(frame['Lõigu sisu'].to_list())
data['algus_lopp']=np.where(data['Lõigu sisu'].isin(alguse_tagid), 'algus', None)
data

lopu_tagid=set(frame2['Lõigu sisu'].to_list())
data['algus_lopp']=np.where(data['Lõigu sisu'].isin(lopu_tagid), 'lopp', data['algus_lopp'])
data

#Lisaks tuli märge panna ka sinna, kus uus kohtulahend algab, sest osades lahendites kohtukulud lahendit ei lõpetanud
kohtuasja_vahetus=set(['Koda'])
kohtuasja_vahetus
data['algus_lopp']=np.where(data['Pealkiri1'].isin(kohtuasja_vahetus), 'vahetus', data['algus_lopp'])

# Samuti tuli märge panna ka sinna kus oli resolutsioon, sest ka see osa on otsingul oluline juristi jaoks
reso_tag=set(['Resolutsioon'])
data['algus_lopp']=np.where(data['Pealkiri1'].isin(reso_tag), 'algus', data['algus_lopp'])

# Vältimaks olukorda, kus kui algust ei leia ja siis võtaks filtreerimisse kogu lahendi,
# filtreeritakse välja ainult need lahendid, kus algus on olemas
# kasutan unikaalsed_ID_d_filter määratlust
Algusega_frame=data[data.Kohtuasja_ID.isin(unikaalsed_ID_d_filter)]
Algusega_frame

# Seejärel tuli ffilliga täita kõik vahepealsed osad ehk algusest kuni "vahetus" või lõpp" sõnani.
Algusega_frame['algus_lopp']=Algusega_frame.algus_lopp.fillna(method='ffill')
Algusega_frame.tail(40)

#Kui filtreerida välja algused saab tabeli, kus otsitavate hulka on vähendatud ja täpsustatud.
filtreeritud_tabel=Algusega_frame[Algusega_frame.algus_lopp=='algus']
filtreeritud_tabel.tail(40)

filtreeritud_tabel.Kohtuasja_ID.nunique() # tulemusena jäi järele 6438 kohtuasja

#Lõin eraldi filtreeritud tabelist csv faili, et ei peaks iga kord hakkama sama protseduuri kordama
filtreeritud_tabel.to_csv("Pohitabel_filt_algusega.csv", index=False)
