==> Mag.too puhas kood/8. Oma CountVectorizer1.ipynb <==
#NB!https://github.com/estnltk/estnltk/blob/master/README.md 
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

#Katsetatud 3 tabeliga:
#Pohitabel_stopid_eemald_lower
#Pohitabel_ainult_reso
#Pohitabel_ilma_resota
data = pd.read_csv("Pohitabel_stopid_eemald_lower.csv")
data

#paneme listi stop_eemaldatud
puhas=data[~pd.isnull(data.stop_eemaldatud)] #eemaldame enne ära null väärtused
column_list=puhas['stop_eemaldatud'].to_list()

# https://colab.research.google.com/drive/1Uh-w_oCllFMaUQOBZVL6xsdSyOfXBTxK?usp=sharing#scrollTo=AUORCu_ztPwG 
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
vectorizer = CountVectorizer()

#Meetod teksti sõnestamiseks ja lemmatiseerimiseks estnltk abil
from estnltk import Text
def lemmatize_with_estnltk(input_string):
    lemmas = []
    text = Text(input_string).tag_layer()
    for word in text.morph_analysis:
        lemmas.append(word.lemma[0])
    return lemmas

vectorizer = TfidfVectorizer(tokenizer = lemmatize_with_estnltk)

# Andmete viimine vektorkujule
matrix = vectorizer.fit_transform(column_list)
matrix.shape

#salvestame matrixi maha, et ei peaks iga kord uuesti laadima
#https://stackoverflow.com/questions/8955448/save-load-scipy-sparse-csr-matrix-in-portable-data-format
#from scipy import sparse

#sparse.save_npz("countvec_matrix.npz", matrix)

#uuesti avamiseks
#your_matrix_back = sparse.load_npz("countvec_matrix.npz")

#loodud proovilause, millega testida sarnasusi
listike=['käibemaksukohustuslaseks registreerimine']

# Proovilause viimine vektorkujule
vordluslist = vectorizer.transform(listike)
vordluslist.shape

#Leian koossiinuse sarnasused
#Leian esimese 200 sarnase indeksid
#Leian esimese 200 koossiinuse sarnasused, mis vastavad indeksitele
#ühendan indeksid ja koosiinuse sarnasused ühte tabelisse
sarnased_maatriks=cosine_similarity(matrix, vordluslist)
sorteeritud_indeksid=np.argsort(np.squeeze(sarnased_maatriks))
listik_indx=[]
for i in range(1,201):
    listik_indx.append(sorteeritud_indeksid[len(sorteeritud_indeksid) - i])
koos_dist=sarnased_maatriks[listik_indx]
koos_dist_jarj=[]
for i in koos_dist:
    koos_dist_jarj.append(i[0])      
print(koos_dist_jarj)
print(listik_indx)
tabel_sarnasus = pd.DataFrame(columns=['Kohtuasja_indx', 'Koos_sarnasus', 'Jarjekord'])
tabel_sarnasus['Kohtuasja_indx'] = listik_indx
tabel_sarnasus['Koos_sarnasus'] = koos_dist_jarj
tabel_sarnasus['Jarjekord'] = list(range(1, 201))
Tabel_sarnased=tabel_sarnasus.set_index('Kohtuasja_indx') 
#tabel_sarnasus
Tabel_sarnased

#Filtreerin indeksite põhjal välja sarnased lõigud
filtreeritud_sarnased=data[data.index.isin(listik_indx)]
filtreeritud_sarnased

# Ühendan kaks tabelit ja filtreerin välja Koos_sarnasus,	Jarjekord, Lõigu sisu, Kohtuasja_nr
Tulemus = pd.concat([Tabel_sarnased, filtreeritud_sarnased], axis=1)
(Tulemus[['Koos_sarnasus', 'Jarjekord','Lõigu sisu', 'Kohtuasja_nr']]).sort_values('Jarjekord', ascending=True)

# Tulemuse salvestamine - 3 tabelit
#Count_vec_tulemus
#Count_vec_tulemus_ainult_reso
#Count_vec_tulemus_ilma_resota
Tulemus.to_csv("Count_vec_tulemus_KMKR.csv", index=False)
