==> Mag.too puhas kood/9. Lemmad.ipynb <==
import pandas as pd
from estnltk import Text

data = pd.read_csv("Pohitabel_stopid_eemald_lower.csv")
data

data_nona=data[~pd.isnull(data.stop_eemaldatud)]
data_nona.shape, data.shape

#Lemmatiseerin kogu tabeli läbi ja kirjutan eraldi faili. Vahepeal kirjutan ka faili juhuks, kui katkestusi peaks olema.
lemmas_all = []
lemmas_id = []
for i, row in data_nona.iterrows():
  text = Text(row['stop_eemaldatud']).tag_layer()
  lemmas = ' '.join([w[0] for w in text.morph_analysis.lemma])
  lemmas_id.append(i)
  lemmas_all.append(lemmas)
  if i%1000==0:
    lemmade_tabel = pd.DataFrame(columns=['Lemma', 'Lemma_id'])
    lemmade_tabel['Lemma'] = lemmas_all
    lemmade_tabel['Lemma_id'] = lemmas_id
    lemmade_tabel.to_csv("Lemmad.csv", index=False)
    print(i)
lemmade_tabel = pd.DataFrame(columns=['Lemma', 'Lemma_id'])
lemmade_tabel['Lemma'] = lemmas_all
lemmade_tabel['Lemma_id'] = lemmas_id
lemmade_tabel.to_csv("Lemmad.csv", index=False)

#Määran lemmade tabeli indeksi Lemma_id, et lihtsustada hilisemat tabelite ühendamist
data2 = pd.read_csv("Lemmad.csv", index_col='Lemma_id')
data2

# Ühendan kaks tabelit
Tulemus = pd.concat([data, data2], axis=1)
Tulemus

#Kirjutan loodud tabeli faili
Tulemus.to_csv('Koondtabel_lemmadega.csv', index=False)

#Filtreerin välja need, millel on reso ja kirjutan faili
Ainult_reso=Tulemus[Tulemus.Pealkiri1=='Resolutsioon']
Ainult_reso.to_csv('Koondtabel_lemmadega_ainult_reso.csv', index=False)

#Filtreerin välja kohtu hinnangu osa ilma resota ja kirjutan faili
Ilma_resota=Tulemus[Tulemus.Pealkiri1!='Resolutsioon']
Ilma_resota.to_csv('Koondtabel_lemmadega_ilma_resota.csv', index=False)
