==> Mag.too puhas kood/10. Word2vec.ipynb <==
# https://radimrehurek.com/gensim/auto_examples/tutorials/run_word2vec.html#sphx-glr-download-auto-examples-tutorials-run-word2vec-py
import pandas as pd
from sklearn.metrics.pairwise import cosine_similarity
import numpy as np

#pip install gensim==3.8.3

#Kolm erinevat varianti:
#Koondtabel_lemmadega.csv
#Koondtabel_lemmadega_ilma_resota.csv
#Koondtabel_lemmadega_ainult_reso.csv
data = pd.read_csv("Koondtabel_lemmadega_ainult_reso.csv")
data

column_list=data['Lemma'].to_list()
column_list[0]

from gensim import utils

class MyCorpus:
    """An iterator that yields sentences (lists of str)."""

    def __iter__(self):
        for x in column_list:
            yield str(x).split()

import gensim.models

sentences = MyCorpus() #lõhub lause tükkideks
model = gensim.models.Word2Vec(sentences=sentences, min_count=0) #treenib mudeli ette antud andmete peal

#Näidis ühe sõnaga
model.wv.most_similar('maks') 

#Näidis nr 2
model.wv.most_similar('käibemaksukohustuslane') 

# funktsioon pikemate lausevektorite loomiseks (sum)
def get_sum_vector(word2vec_model, words):
    words = [word for word in words if word in word2vec_model.wv.vocab]
    if len(words) >= 1:
        return np.sum(word2vec_model[words], axis=0)
    else:
        return []

# Käib kogu mudeli üle, selleks, et saada igast lõigust vektorid kasutades eelnevat funktsiooni get_sum_vector
sonad_arrays=[]
doc_jarjekord = []
for doc in sentences:
  #print(doc)
  vec = get_sum_vector(model, doc)
  sonad_arrays.append(vec)
  doc_jarjekord.append(doc)

#loodud proovilause, millega testida sarnasusi
listike='sisendkäibemaksu mahaarvamine'

#Lemmaks
from estnltk import Text
text = Text(listike).tag_layer()
lemma_list = ' '.join([w[0] for w in text.morph_analysis.lemma])
lemma_list

#teeb ette antud üheelemendilise listi vektoriks
#proovimudel nt get_sum_vector
def proovilistvektoriks(proovimudel, model, lemma_list):
    proovivektor=proovimudel(model, lemma_list.split())
    return proovivektor

proovivektoriks1=proovilistvektoriks(get_sum_vector, model, lemma_list)



#Loon tabeli, kus on koossiinuse sarnaseimad 200 indeksit, sarnasus ja järjekord (kui sarnane on)
sarnased_maatriks=cosine_similarity(sonad_arrays, proovivektoriks1.reshape(1, -1))
sorteeritud_indeksid=np.argsort(np.squeeze(sarnased_maatriks))
listik_indx=[]
for i in range(1,201):
  listik_indx.append(sorteeritud_indeksid[len(sorteeritud_indeksid) - i])
koos_dist=sarnased_maatriks[listik_indx]
koos_dist_jarj=[]
for i in koos_dist:
    koos_dist_jarj.append(i[0]) 
print(koos_dist_jarj)
print(listik_indx)
tabel_sarnasus = pd.DataFrame(columns=['Kohtuasja_indx', 'Koos_sarnasus', 'Jarjekord'])
tabel_sarnasus['Kohtuasja_indx'] = listik_indx
tabel_sarnasus['Koos_sarnasus'] = koos_dist_jarj
tabel_sarnasus['Jarjekord'] = list(range(1, 201))
Tabel_sarnased=tabel_sarnasus.set_index('Kohtuasja_indx') 
Tabel_sarnased


#Filtreerin indeksite põhjal välja sarnased lõigud
filtreeritud_sarnased=data[data.index.isin(listik_indx)]
filtreeritud_sarnased

# Ühendan kaks tabelit ja filtreerin välja Koos_sarnasus,	Jarjekord, Lõigu sisu, Kohtuasja_nr
Tulemus = pd.concat([Tabel_sarnased, filtreeritud_sarnased], axis=1)
Uus = (Tulemus[['Koos_sarnasus', 'Jarjekord','Tag','Lõigu sisu', 'Pealkiri1','Kohtuasja_nr','Kohtuasja_ID', 'algus_lopp', 'Lemma']]).sort_values('Jarjekord', ascending=True)
Uus['Indeks'] = Uus.index
Uus

# Salvestame tulemuse:
#Word2vec_tulemus_lemmatud.csv
#Word2vec_tulemus_ilma_resota_lemmatud.csv
#Word2vec_tulemus_ainult_reso_lemmatud.csv
#Uus.to_csv("Word2vec_tulemus_lemmatud.csv", index=False)
